Add Render.com deployment support
- render.yaml: Render blueprint with free PostgreSQL database
- Dockerfile.render: PHP 8.3 CLI + pdo_pgsql for Render
- start.render.sh: Render-specific startup (artisan serve on $PORT)
- config/database.php: dual MySQL + PostgreSQL support
- Migrations: double() -> decimal() for cross-DB compatibility
- .env.example: documented both MySQL and PostgreSQL configs

+++ database/migrations/2026_07_04_000004_create_services_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
             $table->uuid('company_id');
             $table->string('name');
             $table->text('description')->nullable();
-            $table->double('price', 10, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->useCurrent()->useCurrentOnUpdate();
